userservice/requirements.txt
Django==4.0.2
PyJWT==2.3.0
asgiref==3.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
wheel==0.36.2
psycopg2-binary==2.9.3
gunicorn==20.1.0